Improve tpl

+++ tpl/lib/util.c
@@ -7,13 +7,13 @@
 #include <stdlib.h>
 #include <string.h>
 
-int parse_ints(char *numbers_in, uint64_t **numbers_ptr) {
+size_t parse_ints(char *numbers_in, uint64_t **numbers_ptr) {
     char *numbers_str = strdup(numbers_in);
     char *numbers_str_orig = numbers_str;
     
     uint64_t *numbers = NULL;
     char *token = NULL;
-    int i = 0;
+    size_t i = 0;
     while((token = strtok_r(numbers_str, " ", &numbers_str))) {
         numbers = realloc(numbers, (i + 1) * sizeof(uint64_t));
