+++ tpl/lib/util.h
@@ -1,4 +1,4 @@
 
 #include <stdint.h>
 
-int parse_ints(char *numbers_in, uint64_t **numbers_ptr);
+size_t parse_ints(char *numbers_in, uint64_t **numbers_ptr);
